refactored, only 20 bitstream item are being created
